Paydantic fix + schema gen
