+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/wfReactInterface/generatedBaseObjectClasses.zod.ts
@@ -1,5 +1,13 @@
 import * as z from 'zod';
 
+export const LeaderboardColumnSchema = z.object({
+  evaluation_object_ref: z.string(),
+  scorer_name: z.string(),
+  should_minimize: z.union([z.boolean(), z.null()]).optional(),
+  summary_metric_path_parts: z.array(z.string()).optional(),
+});
+export type LeaderboardColumn = z.infer<typeof LeaderboardColumnSchema>;
+
 export const TestOnlyNestedBaseModelSchema = z.object({
   a: z.number(),
 });
@@ -16,6 +24,13 @@ export type TestOnlyNestedBaseObject = z.infer<
   typeof TestOnlyNestedBaseObjectSchema
 >;
 
+export const LeaderboardSchema = z.object({
+  columns: z.array(LeaderboardColumnSchema),
+  description: z.union([z.null(), z.string()]).optional(),
+  name: z.union([z.null(), z.string()]).optional(),
+});
+export type Leaderboard = z.infer<typeof LeaderboardSchema>;
+
 export const TestOnlyExampleSchema = z.object({
   description: z.union([z.null(), z.string()]).optional(),
   name: z.union([z.null(), z.string()]).optional(),
@@ -26,6 +41,7 @@ export const TestOnlyExampleSchema = z.object({
 export type TestOnlyExample = z.infer<typeof TestOnlyExampleSchema>;
 
 export const GeneratedBaseObjectClassesZodSchema = z.object({
+  Leaderboard: LeaderboardSchema,
   TestOnlyExample: TestOnlyExampleSchema,
   TestOnlyNestedBaseObject: TestOnlyNestedBaseObjectSchema,
 });
