+++ weave/flow/leaderboard_util.py
@@ -1,12 +1,30 @@
+from typing import Any
+
+from pydantic import BaseModel
+
 from weave.trace.refs import OpRef
 from weave.trace.weave_client import WeaveClient, get_ref
 from weave.trace_server.interface.base_object_classes import leaderboard
 from weave.trace_server.trace_server_interface import CallsFilter
 
 
+class LeaderboardModelEvaluationResult(BaseModel):
+    evaluate_call_ref: str
+    value: Any
+
+
+class ModelScoresForColumn(BaseModel):
+    scores: list[LeaderboardModelEvaluationResult]
+
+
+class LeaderboardModelResult(BaseModel):
+    model_ref: str
+    column_scores: list[ModelScoresForColumn]
+
+
 def get_leaderboard_results(
     spec: leaderboard.Leaderboard, client: WeaveClient
-) -> list[leaderboard.LeaderboardModelResult]:
+) -> list[LeaderboardModelResult]:
     entity, project = client._project_id().split("/")
     calls = client.get_calls(
         filter=CallsFilter(
@@ -22,7 +40,7 @@ def get_leaderboard_results(
         )
     )
 
-    res_map: dict[str, leaderboard.LeaderboardModelResult] = {}
+    res_map: dict[str, LeaderboardModelResult] = {}
     for call in calls:
         # Frustrating that we have to get the ref like this. Since the
         # `Call` object auto-derefs the inputs (making a network request),
@@ -37,11 +55,9 @@ def get_leaderboard_results(
             continue
         model_ref_uri = model_ref.uri()
         if model_ref_uri not in res_map:
-            res_map[model_ref_uri] = leaderboard.LeaderboardModelResult(
+            res_map[model_ref_uri] = LeaderboardModelResult(
                 model_ref=model_ref_uri,
-                column_scores=[
-                    leaderboard.ModelScoresForColumn(scores=[]) for _ in spec.columns
-                ],
+                column_scores=[ModelScoresForColumn(scores=[]) for _ in spec.columns],
             )
         for col_idx, c in enumerate(spec.columns):
             eval_obj_ref = get_ref(call.inputs["self"])
@@ -59,7 +75,7 @@ def get_leaderboard_results(
                 else:
                     break
             res_map[model_ref_uri].column_scores[col_idx].scores.append(
-                leaderboard.LeaderboardModelEvaluationResult(
+                LeaderboardModelEvaluationResult(
                     evaluate_call_ref=call_ref_uri, value=val
                 )
             )
