+++ weave/trace_server/interface/base_object_classes/leaderboard.py
@@ -1,8 +1,10 @@
 from dataclasses import field
-from typing import Any, Optional
+from typing import Optional
 
 from pydantic import BaseModel
+
 from weave.trace_server.interface.base_object_classes import base_object_def
+
 
 class LeaderboardColumn(BaseModel):
     evaluation_object_ref: base_object_def.RefStr
@@ -13,17 +15,3 @@ class LeaderboardColumn(BaseModel):
 
 class Leaderboard(base_object_def.BaseObject):
     columns: list[LeaderboardColumn]
-
-
-class LeaderboardModelEvaluationResult(BaseModel):
-    evaluate_call_ref: base_object_def.RefStr
-    value: Any
-
-
-class ModelScoresForColumn(BaseModel):
-    scores: list[LeaderboardModelEvaluationResult]
-
-
-class LeaderboardModelResult(BaseModel):
-    model_ref: str
-    column_scores: list[ModelScoresForColumn]
